Switch buildfarm API to be similar to runfarm API
