+++ deploy/buildtools/buildconfigfile.py
@@ -11,6 +11,7 @@ from buildtools.buildconfig import BuildConfig
 from awstools.awstools import auto_create_bucket, get_snsname_arn
 from buildtools.buildfarm import BuildFarm
 from util.inheritors import inheritors
+from util.deepmerge import deep_merge
 
 # imports needed for python type checking
 from typing import Dict, Optional, List, Set, Type, Any, TYPE_CHECKING
@@ -83,15 +84,19 @@ class BuildConfigFile:
         self.build_ip_set = set()
 
         # retrieve the build host section
+
+        build_farm_defaults_file = global_build_config_file["build_farm"]["base_recipe"]
         build_farm_config_file = None
-        with open(args.buildfarmconfigfile, "r") as yaml_file:
+        with open(build_farm_defaults_file, "r") as yaml_file:
             build_farm_config_file = yaml.safe_load(yaml_file)
 
-        build_farm_name = global_build_config_file["build_farm"]
-        build_farm_conf_dict = build_farm_config_file[build_farm_name]
+        build_farm_type_name = build_farm_config_file["build_farm_type"]
+        build_farm_args = build_farm_config_file["args"]
 
-        build_farm_type_name = build_farm_conf_dict["build_farm_type"]
-        build_farm_args = build_farm_conf_dict["args"]
+        # add the overrides if it exists
+        override_args = global_build_config_file['build_farm'].get('recipe_arg_overrides')
+        if override_args:
+            build_farm_args = deep_merge(build_farm_args, override_args)
 
         build_farm_dispatch_dict = dict([(x.__name__, x) for x in inheritors(BuildFarm)])
