+++ deploy/sample-backup-configs/sample_config_build.yaml
@@ -2,7 +2,15 @@
 # See https://docs.fires.im/en/stable/Advanced-Usage/Manager/Manager-Configuration-Files.html for documentation of all of these params.
 
 # this refers to build farms defined in config_build_farm.yaml
-build_farm: ec2_build_farm
+build_farm:
+  # managerinit replace start
+  base_recipe: build-farm-recipes/aws_ec2.yaml
+  # Uncomment and add args to override defaults.
+  # Arg structure should be identical to the args given
+  # in the base_recipe.
+  #recipe_arg_overrides:
+  #  <ARG>: <OVERRIDE>
+  # managerinit replace end
 
 builds_to_run:
     # this section references builds defined in config_build_recipes.yaml
